fix: make expo-image-picker optional
