+++ package/expo-package/package.json
@@ -37,6 +37,9 @@
     "expo-file-system": {
       "optional": true
     },
+    "expo-image-picker": {
+      "optional": true
+    },
     "expo-sharing": {
       "optional": true
     },
@@ -47,7 +50,6 @@
   "devDependencies": {
     "expo": "^51.0.17",
     "expo-image-manipulator": "^12.0.5",
-    "expo-image-picker": "^15.0.6",
     "expo-media-library": "^16.0.4"
   },
   "scripts": {
