+++ package/expo-package/src/handlers/index.ts
@@ -3,4 +3,3 @@ export * from './getLocalAssetUri';
 export * from './getPhotos';
 export * from './iOS14RefreshGallerySelection';
 export * from './oniOS14GalleryLibrarySelectionChange';
-export * from './takePhoto';
