+++ package/expo-package/src/index.js
@@ -6,7 +6,6 @@ import {
   getPhotos,
   iOS14RefreshGallerySelection,
   oniOS14GalleryLibrarySelectionChange,
-  takePhoto,
 } from './handlers';
 
 import {
@@ -17,6 +16,7 @@ import {
   setClipboardString,
   shareImage,
   Sound,
+  takePhoto,
   triggerHaptic,
   Video,
 } from './optionalDependencies';
